+++ examples/ExactlyOnce/Program.cs
@@ -74,13 +74,13 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             switch (command)
             {
                 case "create_words":
-                    await new WordCreator().Execute(configuration, transactionalId, adminClient, cts);
+                    await WordCreator.Execute(configuration, transactionalId, adminClient, cts);
                     break;
                 case "reverse_words":
-                    await new WordReverser().Execute(configuration, transactionalId, adminClient, cts);
+                    await WordReverser.Execute(configuration, transactionalId, adminClient, cts);
                     break;
                 case "print_words":
-                    await new WordPrinter().Execute(configuration, adminClient, cts);
+                    await WordPrinter.Execute(configuration, adminClient, cts);
                     break;
                 default:
                     logger.LogError("Unknown command");
@@ -88,535 +88,6 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             }
 
             completedCancellationTokenSource.Cancel();
-        }
-    }
-
-    /// <summary>
-    ///     Error that is thrown if the transactional producer receives a
-    ///     fatal error.
-    /// </summary>
-    public class TransactionalProcessorFatalError : Exception
-    {
-    }
-
-    /// <summary>
-    ///     Represents a single message to be sent
-    ///     to an output topic.
-    /// </summary>
-    public class ProduceMessage<K, V>
-    {
-        /// <summary>
-        ///     Name of the topic to produce to.
-        /// </summary>
-        public string Topic { get; set; }
-
-        /// <summary>
-        ///     Message to produce.
-        /// </summary>
-        public Message<K, V> Message { get; set; }
-
-        /// <summary>
-        ///     Delivery handler for this message.
-        /// </summary>
-        public Action<DeliveryReport<K, V>> DeliveryHandler { get; set; } = null;
-    }
-
-    /// <summary>
-    ///     The transactional processor allows to use a producer an consumer for transactions,
-    ///     by recreating the producer when a fatal error happens, sending offsets to transaction
-    ///     and handling aborts and retries correctly.
-    /// </summary>
-    public sealed class TransactionalProcessor<KC, VC, KP, VP> : IDisposable
-    {
-        private static readonly ILogger logger = Utils.LoggerFactory.CreateLogger<TransactionalProcessor<KC, VC, KP, VP>>();
-
-        private IConsumer<KC, VC> consumer { get; set; }
-
-        private IProducer<KP, VP> producer { get; set; }
-
-        private List<ConsumeResult<KC, VC>> inputBatch = new();
-
-        private List<ProduceMessage<KP, VP>> outputBatch = new();
-
-        private bool recreateProducer = true;
-
-        private ConsumerConfig consumerConfig { get; set; }
-
-        private DateTime? lastEndTransactionTime = null;
-
-        private readonly ICollection<ProduceMessage<KP, VP>> EmptyList =
-            new List<ProduceMessage<KP, VP>>(0).AsReadOnly();
-
-        /// <summary>
-        ///     Collection of input topic.
-        /// </summary>
-        public ICollection<string> InputTopics { get; set; }
-
-        /// <summary>
-        ///     Interval between retries.
-        /// </summary>
-        public TimeSpan RetryInterval { get; set; }
-
-        /// <summary>
-        ///     Timeout used for transactional operations.
-        /// </summary>
-        public TimeSpan LocalTransactionOperationTimeout { get; set; }
-
-        /// <summary>
-        ///     Producer config.
-        /// </summary>
-        public ProducerConfig ProducerConfig { get; set; }
-
-        /// <summary>
-        ///     Timeout used for consume calls.
-        /// </summary>
-        public int ConsumeTimeout { get; set; } = 100;
-
-        /// <summary>
-        ///     Max messages to be committed.
-        /// </summary>
-        public int CommitMaxMessages { get; set; } = 1;
-
-        /// <summary>
-        ///     Max period between commits.
-        /// </summary>
-        public TimeSpan CommitPeriod { get; set; } = TimeSpan.MaxValue;
-
-        /// <summary>
-        ///     Recreate the producer on fatal errors like 
-        ///     producer fenced and continue processing.
-        /// </summary>
-        public bool RecreateProducerOnFatalErrors { get; set; } = true;
-
-        /// <summary>
-        ///     Consumer configuration, don't set for a producer only
-        ///     processor.
-        /// </summary>
-        public ConsumerConfig ConsumerConfig
-        {
-            get
-            {
-                return consumerConfig;
-            }
-            set
-            {
-                if (value != null && consumer == null)
-                {
-                    consumerConfig = value;
-
-                    consumerConfig.EnableAutoCommit = false;
-                    consumerConfig.EnableAutoOffsetStore = false;
-                    consumerConfig.IsolationLevel = IsolationLevel.ReadCommitted;
-
-                    consumer = new ConsumerBuilder<KC, VC>(value)
-                        .SetPartitionsAssignedHandler((c, partitions) =>
-                        {
-                            logger.LogDebug("Assigned {Partitions}", partitions);
-                        })
-                        .SetPartitionsLostHandler((c, partitions) =>
-                        {
-                            // When partitions are lost, don't act.
-                            logger.LogWarning("Partitions lost.");
-                        })
-                        .SetPartitionsRevokedHandler((c, partitions) =>
-                        {
-                            // When partitions are being revoked, commit transaction's remaining partitions
-                            // before revoke is completed.
-                            logger.LogInformation("Partitions revoked.");
-                            if (!CancellationTokenSource.IsCancellationRequested)
-                            {
-                                CommitMaybe();
-                            }
-                        })
-                        .Build();
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Cancellation token source to stop the processor.
-        /// </summary>
-        public CancellationTokenSource CancellationTokenSource { get; set; }
-
-        /// <summary>
-        ///     Delegate for Subscribe.
-        /// </summary>
-        public delegate void SubscribeDelegate();
-
-        /// <summary>
-        ///     Delegate for Process.
-        /// </summary>
-        public delegate ICollection<ProduceMessage<KP, VP>> ProcessDelegate(ConsumeResult<KC, VC> consumeResult);
-
-        /// <summary>
-        ///     Delegate for IsCommitNeeded.
-        /// </summary>
-        public delegate bool IsCommitNeededDelegate(ICollection<ConsumeResult<KC, VC>> inputBatch,
-                                                    ICollection<ProduceMessage<KP, VP>> outputBatch);
-
-        /// <summary>
-        ///     Delegate for EndTransaction.
-        /// </summary>
-        public delegate void EndTransactionDelegate(ICollection<ConsumeResult<KC, VC>> inputBatch,
-                                                    ICollection<ProduceMessage<KP, VP>> outputBatch,
-                                                    bool committed);
-
-        /// <summary>
-        ///     Subscribes the consumer to topics.
-        /// </summary>
-        public SubscribeDelegate Subscribe { get; set; }
-
-        /// <summary>
-        ///     Process a new message, or null if no consumer was configured,
-        ///     applies required computation on that data and returns the list
-        ///     of messages to be sent to the output topic.
-        ///     
-        ///     Any exception thrown will cause a rewind of the consumer
-        ///     and an abort of current transaction.
-        /// </summary>
-        public ProcessDelegate Process { get; set; }
-
-        /// <summary>
-        ///     Given current batch of messages received and to be sent to 
-        ///     the output topic, returns true if it's the moment to commit the
-        ///     transaction.
-        /// </summary>
-        public IsCommitNeededDelegate IsCommitNeeded { get; set; }
-
-        /// <summary>
-        ///     Called at the end of transaction with the Kafka batch 
-        ///     and the <paramref name="committed" /> param that says if
-        ///     the transaction completed successfully. Here the application
-        ///     must commit or abort any other transactions it has
-        ///     ongoing.
-        /// </summary>
-        public EndTransactionDelegate EndTransaction { get; set; }
-
-
-        public TransactionalProcessor()
-        {
-            Subscribe = new SubscribeDelegate(DefaultSubscribe);
-            Process = new ProcessDelegate(DefaultProcess);
-            IsCommitNeeded = new IsCommitNeededDelegate(DefaultIsCommitNeeded);
-            EndTransaction = new EndTransactionDelegate(DefaultEndTransaction);
-        }
-
-        private void DefaultSubscribe()
-        {
-            if (consumer != null)
-            {
-                consumer.Subscribe(InputTopics);
-            }
-        }
-
-        private ICollection<ProduceMessage<KP, VP>> DefaultProcess(ConsumeResult<KC, VC> consumeResult)
-        {
-            return EmptyList;
-        }
-
-        private bool DefaultIsCommitNeeded(ICollection<ConsumeResult<KC, VC>> inputBatch,
-                                           ICollection<ProduceMessage<KP, VP>> outputBatch)
-        {
-            return (inputBatch.Any() || outputBatch.Any()) &&
-                (
-                    inputBatch.Count >= CommitMaxMessages ||
-                    outputBatch.Count >= CommitMaxMessages ||
-                    DateTime.UtcNow > lastEndTransactionTime + CommitPeriod
-                );
-        }
-
-        private void DefaultEndTransaction(ICollection<ConsumeResult<KC, VC>> inputBatch,
-                                           ICollection<ProduceMessage<KP, VP>> outputBatch,
-                                           bool committed)
-        {
-        }
-
-        public static void ThrowIfNotRetriable(string operation, KafkaException e, TimeSpan retryInterval)
-        {
-            var txnRequiresAbort = e is KafkaTxnRequiresAbortException;
-            var isRetriable = e is KafkaRetriableException;
-            var isFatal = e.Error.IsFatal;
-            logger.LogError("{Operation} Kafka Exception caught: '{Message}', IsFatal: {isFatal}, TxnRequiresAbort: {TxnRequiresAbort}, IsRetriable: {IsRetriable}",
-                            operation, e.Message, isFatal, txnRequiresAbort, isRetriable);
-            if (isFatal || txnRequiresAbort || !isRetriable)
-            {
-                throw e;
-            }
-            Thread.Sleep(retryInterval);
-        }
-
-        /// <summary>
-        ///     Retry executing <paramref name="action"/> until it succeeds,
-        ///     call <paramref name="onKafkaException"/> if a <see cref="Confluent.Kafka.KafkaException" /> occurs.
-        /// </summary>
-        private void Retry(string operation, Action action, Action<string, KafkaException, TimeSpan> onKafkaException = null)
-        {
-            while (!CancellationTokenSource.IsCancellationRequested)
-            {
-                try
-                {
-                    action();
-                    break;
-                }
-                catch (KafkaException e)
-                {
-                    onKafkaException?.Invoke(operation, e, RetryInterval);
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Commits transaction's produced and consumed messages.
-        ///     Aborts transaction if an abortable exception is thrown,
-        ///     recreates the producer if a fatal exception is thrown or
-        ///     retries each operation that is throwing a retriable exception.
-        /// </summary>
-        public void CommitKafkaTransaction(ICollection<ConsumeResult<KC, VC>> inputBatch,
-                                           ICollection<ProduceMessage<KP, VP>> outputBatch)
-        {
-
-            if (ProducerConfig == null)
-            {
-                if (consumer != null)
-                {
-                    try
-                    {
-                        foreach (var message in inputBatch)
-                        {
-                            Retry("ConsumerStoreOffset", () => consumer.StoreOffset(message),
-                                ThrowIfNotRetriable);
-                        }
-                        Retry("ConsumerCommit", () => consumer.Commit(),
-                                ThrowIfNotRetriable);
-                    }
-                    catch (KafkaException e)
-                    {
-                        throw new KafkaTxnRequiresAbortException(e.Error);
-                    }
-                }
-                return;
-            }
-
-            while (!CancellationTokenSource.IsCancellationRequested)
-            {
-                try
-                {
-                    if (recreateProducer)
-                    {
-                        logger.LogInformation("(Re)creating producer");
-                        if (producer != null) producer.Dispose();
-                        producer = new ProducerBuilder<KP, VP>(ProducerConfig).Build();
-                        producer.InitTransactions(LocalTransactionOperationTimeout);
-                        recreateProducer = false;
-                    }
-
-                    logger.LogInformation("Calling BeginTransaction.");
-                    producer.BeginTransaction();
-
-                    logger.LogInformation("Producing {Count} messages.", outputBatch.Count);
-                    foreach (var message in outputBatch)
-                    {
-                        Retry("Produce", () => producer.Produce(message.Topic,
-                                                                message.Message,
-                                                                message.DeliveryHandler), ThrowIfNotRetriable);
-                    }
-
-                    logger.LogInformation("Producing message completed");
-
-                    if (consumer != null)
-                    {
-                        Retry("SendOffsetsToTransaction", () =>
-                        {
-                            logger.LogInformation("Calling SendOffsetsToTransaction");
-                            producer.SendOffsetsToTransaction(
-                                consumer.Assignment.Select(a => new TopicPartitionOffset(a, consumer.Position(a))),
-                                consumer.ConsumerGroupMetadata,
-                                LocalTransactionOperationTimeout);
-                        }, ThrowIfNotRetriable);
-                        logger.LogInformation("SendOffsetsToTransaction completed");
-                    }
-
-                    Retry("CommitTransaction", () =>
-                    {
-                        logger.LogInformation("calling CommitTransaction");
-                        producer.CommitTransaction();
-                    }, ThrowIfNotRetriable);
-
-                    logger.LogInformation("CommitTransaction completed");
-                    break;
-                }
-                catch (KafkaException e)
-                {
-                    logger.LogError("Kafka Exception caught, aborting transaction, trying again in {TotalSeconds} seconds: '{Message}'",
-                                    RetryInterval.TotalSeconds,
-                                    e.Message);
-                    var TxnRequiresAbort = e is KafkaTxnRequiresAbortException;
-                    if (e.Error.IsFatal)
-                    {
-                        if (!RecreateProducerOnFatalErrors)
-                            throw new TransactionalProcessorFatalError();
-                        recreateProducer = true;
-                    }
-                    else if (TxnRequiresAbort)
-                    {
-                        Retry("AbortTransaction", () =>
-                        {
-                            logger.LogInformation("calling AbortTransaction");
-                            producer.AbortTransaction(LocalTransactionOperationTimeout);
-                        }, (operation, eInner, retryInterval) =>
-                        {
-                            var TxnRequiresAbortErrorInner = eInner is KafkaTxnRequiresAbortException;
-                            var IsRetriableInner = eInner is KafkaRetriableException;
-                            var IsFatalInner = eInner.Error.IsFatal;
-                            logger.LogError("AbortTransaction Kafka Exception caught, trying again in {TotalSeconds} seconds: '{Message}', IsFatal: {IsFatal}, TxnRequiresAbort: {TxnRequiresAbortError}, IsRetriable: {IsRetriable}",
-                                            RetryInterval.TotalSeconds, eInner.Message,
-                                            IsFatalInner, TxnRequiresAbortErrorInner,
-                                            IsRetriableInner);
-                            Thread.Sleep(RetryInterval);
-                            if (!TxnRequiresAbortErrorInner && !IsRetriableInner)
-                            {
-                                if (IsFatalInner)
-                                {
-                                    if (!RecreateProducerOnFatalErrors)
-                                        throw new TransactionalProcessorFatalError();
-                                    recreateProducer = true;
-                                }
-                                // Propagate abort to consumer and application
-                                throw e;
-                            }
-                        });
-                        // Propagate abort to consumer and application
-                        throw;
-                    }
-                    Thread.Sleep(RetryInterval);
-                }
-                catch
-                {
-                    logger.LogError("Caught a different type of exception, this shouldn't happen'");
-                    throw;
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Seeks assigned partitions to last committed offset,
-        ///     or to the earliest offset if no offset was committed yet.
-        /// </summary>
-        private void RewindConsumer(TimeSpan timeout)
-        {
-            foreach (var committedOffset in consumer.Committed(consumer.Assignment, timeout))
-            {
-                var position = committedOffset.Offset < 0
-                               ? Offset.Beginning : committedOffset.Offset;
-                consumer.Seek(new TopicPartitionOffset(committedOffset.TopicPartition, position));
-            }
-        }
-
-        private void ResetTransaction()
-        {
-            lastEndTransactionTime = DateTime.UtcNow;
-            inputBatch.Clear();
-            outputBatch.Clear();
-        }
-
-        /// <summary>
-        ///     Sends and commits accumulated messages, if any.
-        ///     This method should commit any application transaction too.
-        /// </summary>
-        private void CommitMaybe()
-        {
-            if (inputBatch.Any() || outputBatch.Any())
-            {
-                CommitKafkaTransaction(inputBatch, outputBatch);
-                EndTransaction(inputBatch, outputBatch, true);
-                ResetTransaction();
-            }
-        }
-
-        /// <summary>
-        ///     Runs the processor loop, calling process methods.
-        /// </summary>
-        public void RunLoop()
-        {
-            ResetTransaction();
-            Subscribe();
-            while (!CancellationTokenSource.IsCancellationRequested)
-            {
-                try
-                {
-                    ConsumeResult<KC, VC> consumeResult = null;
-                    if (consumer != null)
-                    {
-                        consumeResult = consumer.Consume(ConsumeTimeout);
-                        if (consumeResult != null)
-                        {
-                            inputBatch.Add(consumeResult);
-                        }
-                    }
-                    var messages = Process(consumeResult);
-                    outputBatch.AddRange(messages);
-                    var commitNeeded = IsCommitNeeded(inputBatch, outputBatch);
-                    if (commitNeeded)
-                    {
-                        CommitMaybe();
-                    }
-                }
-                catch (KafkaException e)
-                {
-                    var txnRequiresAbort = e is KafkaTxnRequiresAbortException;
-                    logger.LogError("Consumer KafkaException exception: {Message}, TxnRequiresAbort: {TxnRequiresAbort}",
-                                    e.Message, txnRequiresAbort);
-                    if (txnRequiresAbort)
-                    {
-                        if (consumer != null)
-                        {
-                            Retry("Rewind",
-                                () => RewindConsumer(LocalTransactionOperationTimeout),
-                                (operation, e, retryInterval) =>
-                                {
-                                    Thread.Sleep(RetryInterval);
-                                });
-                        }
-                        EndTransaction(inputBatch, outputBatch, false);
-                        ResetTransaction();
-                        Thread.Sleep(RetryInterval);
-                    }
-                    else
-                    {
-                        // This shouldn't happen
-                        throw;
-                    }
-                }
-                catch (Exception e)
-                {
-                    if (e is TransactionalProcessorFatalError) throw;
-
-                    logger.LogError("Process exception: {Message}",
-                                    e.Message);
-                    // Customer exception requires rewind and abort.
-                    if (consumer != null)
-                    {
-                        Retry("Rewind",
-                            () => RewindConsumer(LocalTransactionOperationTimeout),
-                            (operation, e, retryInterval) =>
-                            {
-                                Thread.Sleep(RetryInterval);
-                            });
-                    }
-                    EndTransaction(inputBatch, outputBatch, false);
-                    ResetTransaction();
-                    Thread.Sleep(RetryInterval);
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Implement IDisposable.
-        /// </summary>
-        public void Dispose()
-        {
-            if (consumer != null) consumer.Close();
-            if (producer != null) producer.Dispose();
         }
     }
 
@@ -637,17 +108,17 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
     ///     Generates random words and sends them to a topic with
     ///     exactly once semantics. This is the producer-only example.
     /// </summary>
-    public class WordCreator
+    public static class WordCreator
     {
-        private readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordCreator));
+        private static readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordCreator));
 
-        private readonly Random random = new();
+        private static readonly Random random = new();
 
-        private readonly List<string> words = new();
+        private static readonly List<string> words = new();
 
-        private WordCreatorConfig config;
+        private static WordCreatorConfig config;
 
-        private string CreateWord()
+        private static string CreateWord()
         {
             int start = (int)'a';
             int end = (int)'z';
@@ -663,15 +134,15 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             return sb.ToString();
         }
 
-        public async Task Execute(IConfiguration configuration, string transactionalId, IAdminClient adminClient, CancellationTokenSource cts)
+        public static async Task Execute(IConfiguration configuration, string transactionalId, IAdminClient adminClient, CancellationTokenSource cts)
         {
             config = configuration.GetSection("WordCreator").Get<WordCreatorConfig>();
             config.Producer.ThrowIfContainsNonUserConfigurable();
             config.Producer.TransactionalId = transactionalId;
             await Utils.CreateTopicMaybe(adminClient, config.OutputTopic.Name, config.OutputTopic.NumPartitions);
 
-            using var transactionalProcessor =
-                    new TransactionalProcessor<Null, string,
+            using var processor =
+                    new Processor<Null, string,
                                                Null, string>
                     {
                         RetryInterval = config.RetryInterval,
@@ -681,11 +152,11 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
                         ProducerConfig = config.Producer,
                         CancellationTokenSource = cts
                     };
-            await Task.Run(() => { transactionalProcessor.RunLoop(); });
+            await Task.Run(() => { processor.RunLoop(); });
             logger.LogInformation("Process finished");
         }
 
-        public List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> _)
+        public static List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> _)
         {
             if (!words.Any())
             {
@@ -703,7 +174,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             return ret;
         }
 
-        public void EndTransaction(ICollection<ConsumeResult<Null, string>> _,
+        public static void EndTransaction(ICollection<ConsumeResult<Null, string>> _,
                                    ICollection<ProduceMessage<Null, string>> _1,
                                    bool committed)
         {
@@ -745,11 +216,11 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
     ///     Read words from a topic, reverse them, and send them to a different topic with
     ///     exactly once semantics. This is the producer and consumer example.
     /// </summary>
-    public class WordReverser
+    public static class WordReverser
     {
-        private readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordReverser));
+        private static readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordReverser));
 
-        private WordReverserConfig config;
+        private static WordReverserConfig config;
 
         private static string Reverse(string original)
         {
@@ -758,7 +229,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             return new string(originalChars);
         }
 
-        public async Task Execute(IConfiguration configuration, string transactionalId, IAdminClient adminClient, CancellationTokenSource cts)
+        public static async Task Execute(IConfiguration configuration, string transactionalId, IAdminClient adminClient, CancellationTokenSource cts)
         {
             config = configuration.GetSection("WordReverser").Get<WordReverserConfig>();
             config.Consumer.ThrowIfContainsNonUserConfigurable();
@@ -769,7 +240,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             await Utils.CreateTopicMaybe(adminClient, config.OutputTopic.Name, config.OutputTopic.NumPartitions);
 
             using var transactionalProcessor =
-                    new TransactionalProcessor<Null, string,
+                    new Processor<Null, string,
                                                Null, string>
                     {
                         RetryInterval = config.RetryInterval,
@@ -808,7 +279,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             logger.LogInformation("Process finished");
         }
 
-        public List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> consumeResult)
+        public static List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> consumeResult)
         {
             var ret = new List<ProduceMessage<Null, string>>();
             if (consumeResult != null && consumeResult.Message != null)
@@ -863,13 +334,13 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
     ///     Reads words from a topic, prints them to standard output with
     ///     exactly once semantics. This is the consumer-only example.
     /// </summary>
-    public class WordPrinter
+    public static class WordPrinter
     {
-        private readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordPrinter));
+        private static readonly ILogger logger = Utils.LoggerFactory.CreateLogger(typeof(WordPrinter));
 
-        private WordPrinterConfig config;
+        private static WordPrinterConfig config;
 
-        private List<string> words = new();
+        private static List<string> words = new();
 
         private static void Print(List<string> words)
         {
@@ -879,7 +350,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             }
         }
 
-        public async Task Execute(IConfiguration configuration, IAdminClient adminClient, CancellationTokenSource cts)
+        public static async Task Execute(IConfiguration configuration, IAdminClient adminClient, CancellationTokenSource cts)
         {
             config = configuration.GetSection("WordPrinter").Get<WordPrinterConfig>();
             config.Consumer.ThrowIfContainsNonUserConfigurable();
@@ -887,7 +358,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             await Utils.CreateTopicMaybe(adminClient, config.InputTopic.Name, config.InputTopic.NumPartitions);
 
             using var transactionalProcessor =
-                    new TransactionalProcessor<Null, string,
+                    new Processor<Null, string,
                                                Null, string>
                     {
                         RetryInterval = config.RetryInterval,
@@ -924,7 +395,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             logger.LogInformation("Process finished");
         }
 
-        public List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> consumeResult)
+        public static List<ProduceMessage<Null, string>> Process(ConsumeResult<Null, string> consumeResult)
         {
             if (consumeResult != null && consumeResult.Message != null)
             {
@@ -933,7 +404,7 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
             return new List<ProduceMessage<Null, string>>();
         }
 
-        public void EndTransaction(ICollection<ConsumeResult<Null, string>> _,
+        public static void EndTransaction(ICollection<ConsumeResult<Null, string>> _,
                                    ICollection<ProduceMessage<Null, string>> _1,
                                    bool committed)
         {
@@ -954,49 +425,6 @@ namespace Confluent.Kafka.Examples.ExactlyOnce
                 //  to a half-committed state.
             }
             words.Clear();
-        }
-    }
-
-    public class Utils
-    {
-        public static readonly ILoggerFactory LoggerFactory = Microsoft.Extensions.Logging.LoggerFactory.Create(builder =>
-            builder.AddSimpleConsole(options =>
-            {
-                options.IncludeScopes = true;
-                options.SingleLine = true;
-                options.TimestampFormat = "yyyy-MM-ddTHH:mm:ss.fffZ ";
-                options.UseUtcTimestamp = true;
-            }));
-
-        private static readonly ILogger logger = Utils.LoggerFactory.CreateLogger("Utils");
-
-        public static async Task CreateTopicMaybe(IAdminClient adminClient, string name, int partitions)
-        {
-            try
-            {
-                await adminClient.CreateTopicsAsync(new List<TopicSpecification>
-                {
-                    new TopicSpecification
-                    {
-                        Name = name,
-                        NumPartitions = partitions
-                    }
-                });
-                logger.LogInformation("Created topic {Name}", name);
-            }
-            catch (CreateTopicsException e)
-            {
-                if (e.Error.Code != ErrorCode.Local_Partial || e.Results.Any(r => r.Error.Code != ErrorCode.TopicAlreadyExists))
-                {
-                    throw e;
-                }
-                logger.LogInformation("Topic '{Name}' already exists.", name);
-            }
-            catch (Exception e)
-            {
-                logger.LogError("Error occurred creating topic '{Name}': {Message}.", name, e.Message);
-                throw;
-            }
         }
     }
 }
